refactor(connectors): rename to pointer
#103
