+++ connectors/bindings.amqp/src/annotation.js
@@ -1,5 +1,5 @@
 'use strict'
 
-const { uris } = require('@toa.io/libraries/connectors')
+const { uris } = require('@toa.io/libraries/pointer')
 
 exports.annotation = uris.construct
