+++ connectors/bindings.amqp/test/annotations.test.js
@@ -3,7 +3,7 @@
 const { generate } = require('randomstring')
 const mock = { uris: { construct: () => generate() }, Pointer: class {} }
 
-jest.mock('@toa.io/libraries/connectors', () => mock)
+jest.mock('@toa.io/libraries/pointer', () => mock)
 const { annotation } = require('../')
 
 it('should export annotations', () => {
