+++ connectors/storages.mongodb/test/annotations.test.js
@@ -4,7 +4,7 @@ const { generate } = require('randomstring')
 
 const mock = { uris: { construct: () => generate() }, Pointer: class {} }
 
-jest.mock('@toa.io/libraries/connectors', () => mock)
+jest.mock('@toa.io/libraries/pointer', () => mock)
 const { annotation } = require('../')
 
 it('should export annotations', () => {
