+++ connectors/storages.mongodb/types/pointer.d.ts
@@ -1,4 +1,4 @@
-import type { Pointer as Base } from '@toa.io/libraries/connectors/types'
+import type { Pointer as Base } from '@toa.io/libraries/pointer/types'
 
 declare namespace toa.mongodb {
