+++ libraries/pointer/src/.deployment/.variables/credentials.js
@@ -4,7 +4,7 @@ const { concat, letters: { up, down } } = require('@toa.io/libraries/generic')
 
 /**
  * @param {string} prefix
- * @param {toa.connectors.URIs} uris
+ * @param {toa.pointer.URIs} uris
  * @returns {toa.deployment.dependency.Variable[]}
  */
 const credentials = (prefix, uris) => {
