+++ libraries/pointer/src/.deployment/.variables/index.js
