+++ libraries/pointer/src/deployment.js
@@ -2,7 +2,7 @@
 
 const declare = require('./.deployment')
 
-/** @type {toa.connectors.Deployment} */
+/** @type {toa.pointer.Deployment} */
 const deployment = (prefix, instances, uris, extensions = undefined) => {
   const proxies = declare.proxies(prefix, instances, uris, extensions)
   const global = declare.variables(prefix, uris)
