+++ libraries/pointer/src/uris/.construct/schema.yaml
