+++ libraries/pointer/src/uris/.construct/validate.js
