+++ libraries/pointer/src/uris/construct.js
@@ -2,7 +2,7 @@
 
 const { normalize, validate } = require('./.construct')
 
-/** @type {toa.connectors.uris.Constructor} */
+/** @type {toa.pointer.uris.Constructor} */
 const construct = (declaration) => {
   const normalized = normalize(declaration)
