+++ libraries/pointer/test/uris.fixtures.js
