+++ libraries/pointer/types/index.d.ts
